requirements.txt
cycler==0.11.0
dataclasses==0.8
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pathlib==1.0.1
Pillow==8.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.1.1
